Refactor VersionResolver and add default target version for BA performance test
